+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## 1.0.3
+## 2.0.0-dev1
 
 - Restricted field access on [PostgreSQLConnection].
 - Connection-level default query timeout.
 - Option to specify timeout for the transaction's `"COMMIT"` query.
+- Optimized byte buffer parsing and construction with `package:buffer`.
+- Hardened codebase with `package:pedantic` and additional lints.
+- Updated codebase to Dart 2.2.
+- `PostgreSQLResult` and `PostgreSQLResultRow` as the return value of a query.
+  - Returned lists are protected with `UnmodifiableListView`. 
 
 ## 1.0.2
 - Add connection queue size